feat: Update npm dependencies and TypeScript version
This commit updates the npm dependencies and TypeScript version in the package.json file. By keeping our dependencies up to date, we ensure that our project is using the latest stable versions of the required packages. Additionally, updating the TypeScript version allows us to take advantage of new language features and improvements.